successfully displaying search results. working on moving results to different container or page
