+++ watch-dog/src/Components/nav/Navbar.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { Input, Menu } from "semantic-ui-react";
 import NavbarDrop from "./NavbarDrop";
-import Search from "./Search"
+import Search from "../search/Search"
 
 export default class Navbar extends Component {
   state = { activeItem: "home" };
